requirements.txt
bchlib
reedsolo
pandas
pytest
numpy
pyldpc
scikit-commpy
mlflow
matplotlib 